refactor(core): isolate local legend layout
